Updating ILP-based schedulers, add benchmark and plotting code. (#4)
* Updating ILP-based scheduelers to use the common interface.

* Pass tasks to ILP schedulers.

* Ported ILP schedulers to the simulator API.

* More fixes and added benchmark data loader.

* Added logging to the ILP schedulers.

* Flake8 fixes.

* Flake8.

* Fixed non-ILP unit tests.

* Fixed all unit tests.

* Ported FIFO scheduler to the new API.

* Removed released_tasks from the scheduler interface.

* Added support for dependencies to the ILP schedulers.

* Verify task dependencies and resource requirements are satisfied.

* Added scheduling horizon to the schedulers.

* Retrive tasks within scheduling horizon.

* Minor edits.

* Fix.

* Reduced usage of format.

* Isorted imports.

* Fixed method that gets schedulable tasks.

* Added slack and start-up delay plots.

* Improved task error messages.

* Use int for time.

* Fixed tests.

* Initialize random seeds.

* Fixes to the ILP schedulers.

* Addressed comments.

* Added preemptive and scheduling_horizon to base scheduler.

* Fixed unit tests.

+++ workload/resources.py
@@ -1,7 +1,7 @@
-import uuid
 import logging
-from copy import copy
+import uuid
 from collections import defaultdict
+from copy import copy
 from functools import total_ordering
 from typing import Mapping, Optional
 
@@ -232,7 +232,7 @@ class Resources(object):
                     f"{task}")
             self.allocate(resource, task, quantity)
 
-    def get_available_quantity(self, resource: Resource) -> int:
+    def get_available_quantity(self, resource: Resource) -> float:
         """Provides the quantity of the available resources of the given type.
 
         If the resource has a specific `id`, then the quantity of that resource
@@ -252,7 +252,7 @@ class Resources(object):
                 resource_quantity += _quantity
         return resource_quantity
 
-    def get_allocated_quantity(self, resource: Resource) -> int:
+    def get_allocated_quantity(self, resource: Resource) -> float:
         """Get the quantity of the given `resource` that has been allocated.
 
         Args:
